move 1469 test to new class
