+++ e2e/gui/testng.xml
@@ -411,4 +411,10 @@
             <class name="com.epam.pipeline.autotests.PipelinesAndStoragesRepositories"/>
         </classes>
     </test>
+
+    <test name="DataStoragesCLITest">
+        <classes>
+            <class name="com.epam.pipeline.autotests.DataStoragesCLITest"/>
+        </classes>
+    </test>
 </suite>
